20250706 i will display the resource mark to screen on mouseover, i've discovered that the resource has a marker set to true , it was supposed to be a number wasn't it? fixed that so if a creature picks up a resource, it should count how many are left, every step it takes towards the base it should put a resource marker equal to how many are left, if the creature stumbles upon additional resources on the way back, it will try to pick them up, count how many remain, and add that to the resource marker number it is laying down. example, creature goes to the resource, takes one, there are 3 left, it retraces its steps back to the base laying a resource marker equal to 3. it gets to the base, drops the resource, and follows its own resource marker back, as it does so it lowers the 3 to an 2 all the way back to the resource, that way if 3 more creatures see this trail, only 2 will bother following it, as they known there are only 2 resources. our creature reaches the resource again, takes one so there are 2 left, this time since it is already discovered, it will not lay a trail, after initial discovery responsibility for editing the trail falls to creatures travelling to the resource - not from. so on the way back it does not interact with the trail, just follows it. suddenly the creature is killed! it drops the resource. later a creature discovers the 2 trail, it reduces it to 1 as it follows it, it then discovers the dead creature and its dropped resource, it picks it up and take it home, the trail is now on 1 up to where that creature died, then 2 from that point onwards, and there are still 2 left at the resource's original location. a creature finds the 1 trail and follows it, reducing it to 0 as it does, it gets to the point where the creature was killed, the trail goes from being 1 to 2 here, the creature decrements the 2 trail to a 1 trail, all the way to the resource, picks one up now there is 1 left. it returns to base doing nothing to the 1 trail, but as it gets to where that creature was killed the trail is now 0, so it reverts to the sort of initial resource discovery state, and records that there is still 1 resource left, it now lays a new trail saying 1 all the way back to the base.
example
there is a resource pool to the north of the base, a creature discovers it, there are 10 resources left there. there is another resource pool north east of the base, a different creature simultaneously discovers it, there are 30 resources there, they both head back to base and their paths converge. the northern creature goes first, it lays a trail saying there are 10 resources left, the north east creature lays a trail of 30 up to the convergence, but then just follows the path back. so the trail from the base will look like it only has 10 resources, but when you get to the fork you discover there are actually 40 in total.
now what should my creatures prioritise? deplete the 10 resource to clear that trail and allow the creatures to lay down the 30 trail? prioritise the 30 resource pool because it's bigger, and they want to get it before the enemy does? should i record how far away the resource is, and prioritise the closest?
either way i should still have creatures follow 0 trails just to double check if there are any resources remaining, as there could be a fork like this example- and that would allow them to rediscover it, or there could be a dead creature holding a resource. only when it is reduced to -1 should they then ignore it.

so to implement that, the creature needs to count how many resources there are after it tries to take one, it should return home, if the tile current resource marker is -1 then it shold be  laying a resource marker equal to the number of resources remaining, else do not edit it. if a creature follows a resource trail then it should decrement it. and creatures should follow trails even if the trail is 0, only when it is -1 do they not follow it

20250707
implemented the above change to the resource marker trails
make it so that when the creature dies it drops the carried resource, so i need to add that resource to the map, which means i need a new method in case 2 resources get dropped on the same tile, they should add together instead of create 2 new resources.

need to decrement the resourcemakrer when you pick up a resource, this has fixed the behaviour of creature 5, which repeatedly visited the depleted resource over and over

this seems to work well.

TODO creatures will stop dead when they come to a player or vehicle, they should treat them like a wall, or attack or push or pick up the player or vehicle, not decided yet