20250708 today i will test the creature movement and it picking up resources, start by making a new map i need to decrement the resourcemarker as i drop off the resource
this all seems to work together quite well so far.

TODO other things i can do next is kill the creatures, i've half made the kill method of the creature, that resets some variables doubtless i've forgotten some.
a killed creature should drop its resources and carried rubble in its trailers if possible, it should mark that position on the map as having one more resource than it used to if possible. it should mark a blood stain on the map that other creatures will react to.
when new creatures are created they will reuse the dead creature's position in the creatures array, so i need to make sure i've reset all the appropriate values.
so i need a kill button for the creature to test it out and i need a make creature button to test out new creature creation.

on discovery of a dead creature the creature should retreat, going back to the creature base, it should also lay a 'warning' trail, any creature that sees this warning trail or dead body should follow it back to the base. or at least go in a different direction, so basically it will oppose going in the direction of the warning trail, unless that warning trail is lower than the warning trail it is on. when at the base it should raise the alarm and the base will produce a warrior creature, the regular, lets call them worker creatures will not follow that warning trail, they will disperse. the warrior will follow the warning trail to the site of the dead body, it will not edit the warning trail as it goes, but it should lay a reinforcing trail, if the reinforcing trail is equal to the warning trail then workers can follow the reinforcing trail to check if everything is ok. when the warrior arrives at the site of the dead creature it will patrol, methodically exploring a set distance from the site. when it starts covering the same ground it will cancel the patrol and return to the base deleting the warning trail and reinforcing trails. if the warrior defeats an enemy it will do an extra patrol until no enemies are found, then it will cancel the patrol. if the warrior is killed that will also count as a dead body, as i said earlier workers can follow the warning/ reinforcing trail so long as the warning number is less than or equal to the reinforcing number, so a worker creature can check on the warrior and report back if it is dead, it will report a higher warning number this time, producing extra warriors, and it will escalate this way until the creature base cannot produce sufficient warriors to take on the threat. produced warriors will need to know how many fellow warriors are nearby so they know to go take on the threat, they should patrol around the base and new warriors should roar so that the nearby warriors can count how many other warriors there are, when they know there are enough they should do a different roar and head off following the warning trail. if in their patrol they encounter a lower warning trail they will follow that. if it never produces enough warriors there will just be a load of warriors patrolling the base, and no workers are even going to visit that warning trail again.

additionally i think it would be helpful for the creatures to display speech bubbles to display to the player what behaviour they are doing, so when they pick up a resource, when they are laying a trail, when they find a dead end, when they drop off a resource, when they edit a trail, when they detect a dead creature, when they are waiting for backup