20250709 working on the kill method of the creature, need to remove the creature index from the map need to have an alive bool, a lot of the times i loop over the creatures array, i need to check if it is alive, i also need a list of dead creatures, so i can reuse them put some of the variable from the constructor for creature into a reset method, to make it easier to reset them when i reuse the creature, needed to add it to the priority array when reused too
this seems to work, if i add a new creature it will add a new creature to the array, but if i kill one, then add one back in, it will reuse that space in the array, seems to work ok, and it hopefully doesn't have any of the old creature's data like proposed pos etc

if a creature finds a resource it will follow resource markers back to the base. but if a creature picks up the first resource and is starting to lay a trail back to the base, but then dies, then the next creature that finds it will pick up the resource and the only resource marker trail will be back to the resource, so it will keep going to the resource even though it already has a resource, the discovering a dead creature logic will have to override this. alternatively i could set a flag on the resource when it is dropped, saying it's newlyDiscovered, if the creature picks up the dropped resource and sees the newly discovered flag, it should not follow the resource trail, it should follow the lowest explored tiles
reference //problem:newlyDiscovered

this seems to work next time sort out the dead creature logic as per yesterdays notes,