20251014 a creature should have a valueOfDiscoveredBloodStain variable, set to -1 initially, if it finds blood it takes on the value of the bloodstain, if this value is therefore more than -1 the creature should return to the base in a similar way to when it finds a resource, it should lay a warningTrail equal to the value of the bloodstain. if a higher bloodstain is encountered on the way back then start laying a warningTrail of that value.
if this creature is killed before it gets back to the base, it should leave a bloodstain equal to this bloodstain it was reporting on plus it's own bloodstain value. that way another creature that stumbles upon it can resume

other workers that encounter the warning trail should follow it back to the base - away from danger

so far i have made it identify bloodstains and record a bloodstain trail, i'm lending the same method as when carrying a resource to get back to the base

this seems to work with one basic test

testing this the warning trail is one space away from the bloodStain, because we only set it when we move, as a one off we should set it when we find the bloodstain --- alternatively maybe i should only detect the bloodstain when the creature is on top of it? that seems ok since i already increased the size of the bloodStain to be a 9 tile square

i have a broken line of creatures moving north in a corridor, i kill the northen most one, the next sees the blood and goes south, the next one sees the blood but carries on north ... i think when we discover the dead body we should delete the memory of our tail, i think that is the problem, yeah that fixed it.

TODO:
i need to do the logic to avoid the warning trail