20251021 so when the creatures see a warning trail, i want them to follow the explored numbers back to the base, unless they see an unexplored tile, in which case they should forget that they are following a warning trail.
so how do i want to do this specifically? if the adjacent tiles have a warning marker, i should set a flag on the creature like warningSeen. when that is true i should try to return to the base. I can turn it to false when i return to base or i find an unexplored tile.

made a warningInAdjacent5 in the creature method which returns true if there is a warning nearby, i created a bool for the creature called seenWarningBool this is set to true if there is a warning nearby. if that is set to true we will follow the same logic as when we have a resource or found blood - which means we are trying to return to base, but i will set the bool to false again if we are adjacent to an unexplored tile - that is accessible or we return to base

need to test that out, probably need a new map