20251022 would it be noticably more efficient to only check if the creature is on the warning marker, rather than on or adjacent to? would it be better to have the warning marker appear similar to the bloodstain in that it is painted onto 9 tiles instead of 1? that way the creature would still detect being adjacent to the actual warning marker. or would the result even be much different if the creatures only reacted to warnings when they were directly on them?
there is an odd behaviour when setting the warning trail, the creature does not quite return to the base, the explored number on the base is 0, north one from there is 1, and north from there is for some reason also 1, it should be 2, investigating...
it was because when we are next to the base we reset the explored number to 0, change this to reset it to 1

this seems to work quite well, if one creature dies and another discovers it it retreats to the base laying a warning trail as it goes, then creatures will not follow that path anymore

TODO
so next i need to have a warrior creature be produced at the base when the base is warned.
it will be produced and then wait near the creature base until there are enough warriors produced such that their strength is greater than the warning blood stain value
this group will follow the warning path laying a reinforcing trail as they go, equal to their strength.
    as the reinforcing trail will exceed or be equal to the warning trail then the worker creatures can consider the warning dealt with and no longer be scared of the warning, so some of them might even follow
when the warriors follow the warning trail they should arrive at a blood stain related to the warning trail,
    here they could explore, patrol, or wait for a threat to reveal itself.
    if enough time passes or if they successfully defeat a threat they should consume the blood stain deleting it from the map and follow their reinforcing trail to return to the base.
    they will delete the reinforcing trail and warning trail as they go
although perhaps they will not arrive at the blood stain, maybe the player or some other creature has interacted with the warning trail or the blood stain or the terrain like walls to impede the warriors. in such an instance the warriors will basically retreat, explore, patrol or perhaps garrison at the end of the trail, waiting for a threat.