20251103 testing some of the movement with the blood, some creatures walk right through the blood, i think it's because the previous creature got in its way
creature 0 discovers a resource, it turns back, it is about to overtake creature 1 when creature 0 is killed, creature 1 takes creature 0's resource and heads home, creature 2 overtakes creature 1, it is now on the space that creature 0 was killed, it is aware of the blood so it wants to head back, but my head back logic checks for a neighbouring resource and goes there, so it carries on going north since it is right next to the resource. since creature 0 was marked as the first creature to locate that resource it was following its own explored number trail back to base when it was killed, creature 2 technically did not discover it, it was already discovered, so the newly discovered flag is still false for creature 2, so it tries to follow the resource markers back, but they don't go back to teh base, there are only 2 resource markers, so the creature 2 just goes back and forth.
instead of this newly discovered thing, could i just see if there is only one resource marker then go to the lowest explored number that is not that resource marker? what happens if 2 resource marker trail are being laid and they intersect? i think part of the problem is that when a creature is killed carrying a resource, it leaves a resource marker of 1, when it should leave a resource marker equal to the number of resources it discovered. no that is pointless because the creature that picks up that dropped resoruce should then lay a marker equal to the number of resources that are left there which is probably just 0, but i already thought about this, workers will follow that 0 resource trail because it is not -1, and they will discover that next to the 0 trail there is a trail of 9 or whatever, and they will find the original resource pool. my problem is that i am following resource markers back to the base, when i should only be following the lowest explored path back to the base, keep the resource markers as something that you only follow TO the resource not AWAY from the resource.

ok so i deleted the logic that tries to follow a resource marker back to the base, now creatures will see blood and immediately run away, they will not pick up the resource that the dead creature dropped, unless it has to move past it anyway, i think this works better.

TODO: I think it would be useful for the workers to have a step counter when they are exploring, if they go too far they should retunr to base to reset the step counter. also the creature base should keep track of if a creature has not visited in a while and then it can assume it is dead and make more. also the workers should record if they have seen an unexplored tile since the last visit to the base, if the base detects that no resources have been collected in a while and no unexplored tiles are being explored then it can decide to make a different creature capable of seeing throgh walls, this could then try and dig through to new areas.